Ch.7 Benefits of P2SH to Segregated Witness Output and Transaction Examples.
Please review when BIP-68 was implemented on litecoin.  Was it October 2016?
```
BIP-68 and BIP-112 were activated in May 2016 as a soft fork upgrade to the consensus rules on bitcoin and October 2016 for litecoin. 
```
Please also review the technical details in this specific section.  I mentioned RBF was not included in litecoin:
```
===== nSequence as a consensus-enforced relative timelock
```